Sound fix and blink fix
Wasn't happy with the volume I was getting from a passive piezo.
Figured out that if I use active and switch from tone() to
digitalWrite() the volume is much louder.  Changed logic to use the
active piezo with digitalWrite().

This required turning the sound on and back off since I can't send a
time to digitalWrite() and I can't sleep().

I also fixed the blinking when I was setting the alarm.  I needed to
move the printing of the alarm time outside of the loop.