Added backup battery handling, refactoring
The clock now supports a backup battery.  If vin goes down, power
switches over to the battery, if present, and a symbol is printed.
Also, the battery level is checked every hour.  If it goes below 1.5V, a
warning is printed.

If power switches to battery, the brightness is set to the lowest and
separator stops blinking.

Refactored code some more.  Fixed some line wrapping.  Moved functions
to overarching sections to help find them faster.